feat : ScoreReport 리팩토링
전략패턴 적용
