+++ EndGameReport.java
@@ -11,9 +11,8 @@ import java.awt.event.*;
 import javax.swing.*;
 import javax.swing.border.*;
 import javax.swing.event.*;
-
 import java.util.*;
-import java.text.*;
+
 
 public class EndGameReport {
 
@@ -146,4 +145,3 @@ public class EndGameReport {
 		}
 	}
 }
-
