+++ ScoreReportSender.java
@@ -1,0 +1,3 @@
+public interface ScoreReportSender {
+    void send(String recipient, String content);
+}
